+++ lib/unused_code_analyzer.dart
@@ -1,0 +1,8 @@
+export 'package:dart_code_metrics/src/analyzers/unused_code_analyzer/models/unused_code_file_report.dart';
+export 'package:dart_code_metrics/src/analyzers/unused_code_analyzer/models/unused_code_issue.dart';
+
+export 'package:dart_code_metrics/src/analyzers/unused_code_analyzer/reporters/reporters_list/console/unused_code_console_reporter.dart';
+export 'package:dart_code_metrics/src/analyzers/unused_code_analyzer/reporters/reporters_list/json/unused_code_json_reporter.dart';
+
+export 'package:dart_code_metrics/src/analyzers/unused_code_analyzer/unused_code_analyzer.dart';
+export 'package:dart_code_metrics/src/analyzers/unused_code_analyzer/unused_code_config.dart';