+++ test/resources/unused_code_analyzer/export.dart
@@ -1,0 +1,1 @@
+export 'exported_members.dart';