+++ test/resources/unused_code_analyzer/not_used.dart
@@ -1,0 +1,5 @@
+class NotUsed {} // LINT
+
+const int value = 15; // LINT
+
+void someFunction() {} // LINT