+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * fix: ignore private variables in `avoid-global-state` rule.
 * feat: support excludes for a separate anti-pattern.
 * chore: restrict `analyzer` version to `>=2.4.0 <3.2.0`.
+* feat: add check unused code command.
 * feat: ignore private members for `check-unused-l10n` command.
 * chore: remove `@immutable` annotation from code.